backend/pyproject.toml
[tool.poetry]
name = "laptop-review"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "laptop_review"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
firebase = "^4.0.1"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
firebase-admin = "^6.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
